MainControl.class now has method switchView(JPanel p).
MainControl.class switches its view very differently now.
Before, main panels would be added and removed from a
containing panel. Now everything that can be displayed is
added, but its visibility is only true if you switch to 
that view. I think this might use more memory in exchange
for switching speed. I also prefer it as the code is 
cleaner, not requiring calls to revalidate() or repaint().