Added ContactLookupTable.class and FieldMap.class. Code organization.
ContactLookupTable holds a FieldMap for each Field in a Contact object.
A field map is a HashMap<Field, List<UUID>> where the Field value is the
key, and the value is a list of UUIDs that have that value. This makes
it simple to, for example, retrieve all contacts that have the name
"foo", or in the group "bar" etc.

I think a db would be a much better idea at this point. Looking into
sqlite.